BUGFIX: delete from published site never calls canDeleteFromLive(). (via marcus #5364)
git-svn-id: svn://svn.silverstripe.com/silverstripe/open/modules/sapphire/branches/2.4@102386 467b73ca-7a2a-4603-9d3b-597d59a354a9

+++ core/model/SiteTree.php
@@ -1940,7 +1940,7 @@ class SiteTree extends DataObject implements PermissionProvider,i18nEntityProvid
 				if($this->ExistsOnLive) {
 					// "restore"
 					$actions->push(new FormAction('revert',_t('CMSMain.RESTORE','Restore')));
-					if($this->canDelete()) {
+					if($this->canDelete() && $this->canDeleteFromLive()) {
 						// "delete from live"
 						$actions->push(new FormAction('deletefromlive',_t('CMSMain.DELETEFP','Delete from the published site')));
 					}
